+++ feature/dnd/src/main/kotlin/pl/krystiankaniowski/performance/dnd/DndModule.kt
@@ -7,16 +7,16 @@ import dagger.hilt.components.SingletonComponent
 import dagger.multibindings.IntoSet
 import pl.krystiankaniowski.performance.dnd.settings.DndSettingsProvider
 import pl.krystiankaniowski.performance.dnd.timer.DndTimerObserver
+import pl.krystiankaniowski.performance.dnd.usecase.IsDoNotDisturbEnabledUseCase
 import pl.krystiankaniowski.performance.dnd.usecase.IsDoNotDisturbEnabledUseCaseImpl
+import pl.krystiankaniowski.performance.dnd.usecase.SetDoNotDisturbEnabledUseCase
 import pl.krystiankaniowski.performance.dnd.usecase.SetDoNotDisturbEnabledUseCaseImpl
+import pl.krystiankaniowski.performance.dnd.usecase.TurnOffDoNotDisturbUseCase
 import pl.krystiankaniowski.performance.dnd.usecase.TurnOffDoNotDisturbUseCaseImpl
+import pl.krystiankaniowski.performance.dnd.usecase.TurnOnDoNotDisturbUseCase
 import pl.krystiankaniowski.performance.dnd.usecase.TurnOnDoNotDisturbUseCaseImpl
 import pl.krystiankaniowski.performance.domain.settings.SettingsItemsProvider
 import pl.krystiankaniowski.performance.domain.timer.TimerObserver
-import pl.krystiankaniowski.performance.domain.usecase.dnd.IsDoNotDisturbEnabledUseCase
-import pl.krystiankaniowski.performance.domain.usecase.dnd.SetDoNotDisturbEnabledUseCase
-import pl.krystiankaniowski.performance.domain.usecase.dnd.TurnOffDoNotDisturbUseCase
-import pl.krystiankaniowski.performance.domain.usecase.dnd.TurnOnDoNotDisturbUseCase
 
 @Module
 @InstallIn(SingletonComponent::class)
