+++ feature/dnd/src/main/kotlin/pl/krystiankaniowski/performance/dnd/settings/DndSettingsProvider.kt
@@ -4,14 +4,14 @@ import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 import pl.krystiankaniowski.performance.dnd.R
+import pl.krystiankaniowski.performance.dnd.usecase.IsDoNotDisturbEnabledUseCase
+import pl.krystiankaniowski.performance.dnd.usecase.SetDoNotDisturbEnabledUseCase
 import pl.krystiankaniowski.performance.domain.navigation.Destination
 import pl.krystiankaniowski.performance.domain.navigation.Navigator
 import pl.krystiankaniowski.performance.domain.provider.StringsProvider
 import pl.krystiankaniowski.performance.domain.settings.SettingsItem
 import pl.krystiankaniowski.performance.domain.settings.SettingsItemsProvider
 import pl.krystiankaniowski.performance.domain.settings.SettingsOrder
-import pl.krystiankaniowski.performance.domain.usecase.dnd.IsDoNotDisturbEnabledUseCase
-import pl.krystiankaniowski.performance.domain.usecase.dnd.SetDoNotDisturbEnabledUseCase
 import javax.inject.Inject
 
 class DndSettingsProvider @Inject constructor(
