+++ feature/dnd/src/main/kotlin/pl/krystiankaniowski/performance/dnd/usecase/IsDoNotDisturbEnabledUseCaseImpl.kt
@@ -3,7 +3,6 @@ package pl.krystiankaniowski.performance.dnd.usecase
 import kotlinx.coroutines.flow.first
 import pl.krystiankaniowski.performance.dnd.PreferencesKeys
 import pl.krystiankaniowski.performance.domain.repository.AppSettingsRepository
-import pl.krystiankaniowski.performance.domain.usecase.dnd.IsDoNotDisturbEnabledUseCase
 import javax.inject.Inject
 
 class IsDoNotDisturbEnabledUseCaseImpl @Inject constructor(
