+++ feature/dnd/src/main/kotlin/pl/krystiankaniowski/performance/dnd/usecase/SetDoNotDisturbEnabledUseCase.kt
@@ -1,4 +1,4 @@
-package pl.krystiankaniowski.performance.domain.usecase.dnd
+package pl.krystiankaniowski.performance.dnd.usecase
 
 interface SetDoNotDisturbEnabledUseCase {
     suspend operator fun invoke(value: Boolean)
