+++ feature/dnd/src/main/kotlin/pl/krystiankaniowski/performance/dnd/usecase/TurnOffDoNotDisturbUseCaseImpl.kt
@@ -4,7 +4,6 @@ import android.app.NotificationManager
 import android.content.Context
 import androidx.core.app.ActivityCompat.getSystemService
 import dagger.hilt.android.qualifiers.ApplicationContext
-import pl.krystiankaniowski.performance.domain.usecase.dnd.TurnOffDoNotDisturbUseCase
 import javax.inject.Inject
 
 class TurnOffDoNotDisturbUseCaseImpl @Inject constructor(
