+++ feature/dnd/src/main/kotlin/pl/krystiankaniowski/performance/dnd/usecase/TurnOnDoNotDisturbUseCase.kt
@@ -1,4 +1,4 @@
-package pl.krystiankaniowski.performance.domain.usecase.dnd
+package pl.krystiankaniowski.performance.dnd.usecase
 
 interface TurnOnDoNotDisturbUseCase {
     operator fun invoke()
