+++ feature/dnd/src/test/kotlin/pl/krystiankaniowski/performance/dnd/settings/DndSettingsProviderTest.kt
@@ -6,11 +6,11 @@ import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
+import pl.krystiankaniowski.performance.dnd.usecase.IsDoNotDisturbEnabledUseCase
+import pl.krystiankaniowski.performance.dnd.usecase.SetDoNotDisturbEnabledUseCase
 import pl.krystiankaniowski.performance.domain.navigation.Navigator
 import pl.krystiankaniowski.performance.domain.provider.StringsProvider
 import pl.krystiankaniowski.performance.domain.settings.SettingsItem
-import pl.krystiankaniowski.performance.domain.usecase.dnd.IsDoNotDisturbEnabledUseCase
-import pl.krystiankaniowski.performance.domain.usecase.dnd.SetDoNotDisturbEnabledUseCase
 import pl.krystiankaniowski.performance.testing.rule.InstantDispatcherExtension
 
 @ExtendWith(InstantDispatcherExtension::class)
