+++ feature/notifications/src/main/kotlin/pl/krystiankaniowski/performance/notification/NotificationModule.kt
@@ -6,12 +6,14 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import dagger.multibindings.IntoSet
 import pl.krystiankaniowski.performance.domain.Initializer
-import pl.krystiankaniowski.performance.domain.timer.TimerObserver
 import pl.krystiankaniowski.performance.domain.settings.SettingsItemsProvider
-import pl.krystiankaniowski.performance.domain.usecase.notification.StartForegroundServiceUseCase
-import pl.krystiankaniowski.performance.domain.usecase.notification.StopForegroundServiceUseCase
-import pl.krystiankaniowski.performance.notification.timer.NotificationTimerObserver
+import pl.krystiankaniowski.performance.domain.timer.TimerObserver
 import pl.krystiankaniowski.performance.notification.settings.NotificationSettingsProvider
+import pl.krystiankaniowski.performance.notification.timer.NotificationTimerObserver
+import pl.krystiankaniowski.performance.notification.usecase.StartForegroundServiceUseCase
+import pl.krystiankaniowski.performance.notification.usecase.StartForegroundServiceUseCaseImpl
+import pl.krystiankaniowski.performance.notification.usecase.StopForegroundServiceUseCase
+import pl.krystiankaniowski.performance.notification.usecase.StopForegroundServiceUseCaseImpl
 
 @Module
 @InstallIn(SingletonComponent::class)
